# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PostUserOutput": {
				"properties": {
					"accessToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken"
				],
				"type": "object"
			},
			"PostUserInput": {
				"properties": {
					"password": {
						"type": "string"
					},
					"login": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"login"
				],
				"type": "object"
			},
			"Book": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					},
					"author": {
						"type": "string"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"title",
					"summary",
					"author",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookIdParam": {
				"type": "string"
			},
			"CreateBookInput": {
				"properties": {
					"title": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					},
					"author": {
						"type": "string"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"summary",
					"author",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "node-clean-archi",
		"contact": {}
	},
	"paths": {
		"/users/signin": {
			"post": {
				"operationId": "Signin",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostUserOutput"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					},
					"404": {
						"description": "User not found"
					}
				},
				"tags": [
					"users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostUserInput"
							}
						}
					}
				}
			}
		},
		"/users/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostUserOutput"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					},
					"409": {
						"description": "Already exists"
					}
				},
				"tags": [
					"users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostUserInput"
							}
						}
					}
				}
			}
		},
		"/books": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Book"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get all books",
				"tags": [
					"books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Book"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					}
				},
				"summary": "Create a new book",
				"tags": [
					"books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- Book details",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookInput",
								"description": "- Book details"
							}
						}
					}
				}
			}
		},
		"/books/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Book"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					},
					"404": {
						"description": "Book not found"
					}
				},
				"summary": "Get a book by ID",
				"tags": [
					"books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- Book ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/BookIdParam"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": ""
					},
					"400": {
						"description": "Invalid request params"
					},
					"404": {
						"description": "Book not found"
					}
				},
				"summary": "Delete a book by ID",
				"tags": [
					"books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- Book ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/BookIdParam"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8000",
			"description": "Localhost"
		}
	]
}